Add files via upload
The two uploaded files are Salmon Game and the game control.
The game includes a presentation and a salmon game which allows the user to move the fish around and answer certain questions when reaching certain positions. 
The game control maps several keys from the keyboard to a joystick and four push buttons.
It's noticeable that MATLAB isn't the most efficient programming software for game, which the game would crash sometimes unexpectedly. To solve this problem, different combinations of pushbuttons were coded to either restart the game or reset the entire demonstration.